Fix review comments for fetch backend URL
