+++ system/lib/wasmfs/backends/fetch_backend.cpp
@@ -13,13 +13,12 @@
 // See library_wasmfs_fetch.js
 
 extern "C" {
-void _wasmfs_create_fetch_backend_js(wasmfs::backend_t, char*);
+void _wasmfs_create_fetch_backend_js(wasmfs::backend_t, const char*);
 }
 
 namespace wasmfs {
 
-extern "C" backend_t wasmfs_create_fetch_backend(char* base_url) {
-  // TODO: use base url, cache on JS side
+extern "C" backend_t wasmfs_create_fetch_backend(const char* base_url) {
   return wasmFS.addBackend(
     std::make_unique<ProxiedAsyncJSBackend>([base_url](backend_t backend) {
       _wasmfs_create_fetch_backend_js(backend, base_url);
